Moved SQL context to a new folder.
