+++ EyeCT4Rails/EyeCT4RailsDatabase/EyeCT4RailsDatabase.csproj
@@ -44,19 +44,19 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="CleanupSQLContext.cs" />
+    <Compile Include="SQLContexts\CleanupSQLContext.cs" />
     <Compile Include="Database.cs" />
     <Compile Include="DatabaseConfiguration.cs" />
-    <Compile Include="DepotManagementSQLContext.cs" />
-    <Compile Include="MaintenanceSQLContext.cs" />
+    <Compile Include="SQLContexts\DepotManagementSQLContext.cs" />
+    <Compile Include="SQLContexts\MaintenanceSQLContext.cs" />
     <Compile Include="Models\ICleanupContext.cs" />
     <Compile Include="Models\IDepotManagementContext.cs" />
     <Compile Include="Models\IRideManagementContext.cs" />
     <Compile Include="Models\IMaintenanceContext.cs" />
     <Compile Include="Models\IUserContext.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="RideManagementSQLContext.cs" />
-    <Compile Include="UserSQLContext.cs" />
+    <Compile Include="SQLContexts\RideManagementSQLContext.cs" />
+    <Compile Include="SQLContexts\UserSQLContext.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\EyeCT4RailsLib\EyeCT4RailsLib.csproj">
