+++ EyeCT4Rails/EyeCT4RailsDatabase/SQLContexts/MaintenanceSQLContext.cs
@@ -6,7 +6,7 @@ using EyeCT4RailsLib.Classes;
 using EyeCT4RailsLib.Enums;
 using Oracle.ManagedDataAccess.Client;
 
-namespace EyeCT4RailsDatabase
+namespace EyeCT4RailsDatabase.SQLContexts
 {
     public class MaintenanceSqlContext : IMaintenanceContext
     {
