+++ EyeCT4Rails/EyeCT4RailsDatabase/SQLContexts/UserSQLContext.cs
@@ -1,12 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
 using EyeCT4RailsDatabase.Models;
+using EyeCT4RailsLib;
 using EyeCT4RailsLib.Classes;
 using EyeCT4RailsLib.Enums;
 using Oracle.ManagedDataAccess.Client;
 using Oracle.ManagedDataAccess.Types;
 
-namespace EyeCT4RailsDatabase
+namespace EyeCT4RailsDatabase.SQLContexts
 {
     public class UserSqlContext : IUserContext
     {
