+++ EyeCT4Rails/EyeCT4RailsLogic/Repositories/MaintenanceRepository.cs
@@ -1,8 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
-using EyeCT4RailsDatabase;
 using EyeCT4RailsDatabase.Models;
-using EyeCT4RailsLib;
+using EyeCT4RailsDatabase.SQLContexts;
 using EyeCT4RailsLib.Classes;
 using EyeCT4RailsLib.Enums;
 using EyeCT4RailsLogic.Exceptions;
