+++ EyeCT4Rails/EyeCT4RailsLogic/Repositories/RideManagementRepository.cs
@@ -1,8 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using EyeCT4RailsDatabase;
 using EyeCT4RailsDatabase.Models;
+using EyeCT4RailsDatabase.SQLContexts;
 using EyeCT4RailsLib;
 using EyeCT4RailsLib.Classes;
 using EyeCT4RailsLib.Enums;
