+++ EyeCT4Rails/EyeCT4RailsLogic/Repositories/UserRepository.cs
@@ -1,7 +1,6 @@
 ﻿using System;
-using EyeCT4RailsDatabase;
 using EyeCT4RailsDatabase.Models;
-using EyeCT4RailsLib;
+using EyeCT4RailsDatabase.SQLContexts;
 using EyeCT4RailsLib.Classes;
 using EyeCT4RailsLib.Enums;
 using EyeCT4RailsLogic.Exceptions;
